return puma worker

+++ db/migrate/20151226120426_create_points.rb
@@ -1,8 +1,8 @@
 class CreatePoints < ActiveRecord::Migration
   def change
     create_table :points do |t|
-      t.decimal :lng
-      t.decimal :lat
+      t.float :lng
+      t.float :lat
       t.references :point_of_interest, index: true, foreign_key: true
 
       t.timestamps null: false
