+++ db/migrate/20151226201737_add_lng_inc_to_points.rb
@@ -1,5 +1,5 @@
 class AddLngIncToPoints < ActiveRecord::Migration
   def change
-    add_column :points, :lng_inc, :decimal
+    add_column :points, :lng_inc, :float
   end
 end
