+++ db/migrate/20151226201801_add_lat_inc_to_points.rb
@@ -1,5 +1,5 @@
 class AddLatIncToPoints < ActiveRecord::Migration
   def change
-    add_column :points, :lat_inc, :decimal
+    add_column :points, :lat_inc, :float
   end
 end
