+++ db/migrate/20151226203143_add_inc_value_to_points.rb
@@ -1,5 +1,5 @@
 class AddIncValueToPoints < ActiveRecord::Migration
   def change
-    add_column :points, :inc_value, :decimal
+    add_column :points, :inc_value, :float
   end
 end
